+++ padogrid-deployment/src/main/resources/common/bin_sh/.padogrid_completion.bash
@@ -231,14 +231,7 @@ __padogrid_complete()
    -type)
       if [ "$command" == "create_pod" ]; then
          type_list="local vagrant"
-      elif [ "$command" == "create_cluster" ]; then
-         case "$PRODUCT" in
-         geode)
-            type_list="default pado padolite";;
-         *)
-            type_list="default";;
-         esac
-      elif [ "$command" == "make_cluster" ] || [ "$command" == "create_group" ]; then
+      elif [ "$command" == "make_cluster" ] || [ "$command" == "create_cluster" ] || [ "$command" == "create_group" ]; then
          type_list="default"
          local product=""
          for i in $(seq 1 $len); do
@@ -287,7 +280,7 @@ __padogrid_complete()
    -product)
       if [ "$command" == "show_bundle" ]; then
          type_list="$BUNDLE_PRODUCT_LIST"
-      elif [ "$command" == "make_cluster" ] || [ "$command" == "add_cluster" ] || [ "$command" == "create_group" ]; then
+      elif [ "$command" == "make_cluster" ] || [ "$command" == "create_cluster" ] || [ "$command" == "add_cluster" ] || [ "$command" == "create_group" ]; then
          type_list=$(getInstalledProducts)
       elif [ "$command" == "create_docker" ]; then
          type_list="$DOCKER_PRODUCT_LIST"
@@ -532,7 +525,7 @@ __padogrid_complete()
          is_path="true"
       elif [ "$command" == "vc_start" ] &&  [[ "$cur_word" != "-"* ]] && [ $len -gt 3 ]; then
          is_path="true"
-      elif [ "$command" = "make_cluster" ]; then
+      elif [ "$command" = "make_cluster" ] || [ "$command" = "create_cluster" ]; then
          # If -product specified then get the product's app options
          __getArrayElementIndex "-product" "${COMP_WORDS[@]}"
          local index=$?
@@ -1043,14 +1036,7 @@ __command_complete()
    -type)
       if [ "$command" == "create_pod" ]; then
          type_list="local vagrant"
-      elif [ "$command" == "create_cluster" ]; then
-         case "$PRODUCT" in
-         geode)
-            type_list="default pado padolite";;
-         *)
-            type_list="default";;
-         esac
-      elif [ "$command" == "make_cluster" ] || [ "$command" == "create_group" ]; then
+      elif [ "$command" == "make_cluster" ] || [ "$command" == "create_cluster" ] || [ "$command" == "create_group" ]; then
          type_list="default"
          local product=""
          for i in $(seq 1 $len); do
@@ -1071,7 +1057,7 @@ __command_complete()
    -product)
       if [ "$command" == "show_bundle" ]; then
          type_list="$BUNDLE_PRODUCT_LIST"
-      elif [ "$command" == "make_cluster" ] || [ "$command" == "add_cluster" ] || [ "$command" == "create_group" ]; then
+      elif [ "$command" == "make_cluster" ] || [ "$command" = "create_cluster" ] || [ "$command" == "add_cluster" ] || [ "$command" == "create_group" ]; then
          type_list=$(getInstalledProducts)
       elif [ "$command" == "create_docker" ]; then
          type_list="$DOCKER_PRODUCT_LIST"
@@ -1315,7 +1301,7 @@ __command_complete()
          is_path="true"
       elif [ "$command" == "vc_start" ] &&  [[ "$cur_word" != "-"* ]] && [ $len -gt 2 ]; then
          is_path="true"
-      elif [ "$command" = "make_cluster" ]; then
+      elif [ "$command" = "make_cluster" ] || [ "$command" = "create_cluster" ]; then
          # If -product specified then get the product's app options
          __getArrayElementIndex "-product" "${COMP_WORDS[@]}"
          local index=$?